Slave Cumulative: LPTim
Enable LPTim Update DwTick with FIX LpTim Bug with FreeRtos Stm32Duino (Inc HalTick with suppressed Tick Count count difference).
Syncro Publish HeartBeat From Master with Slave also from PowerDown.
Fix bug variables and logic WIND Elaboration.
MASTER: Remove RPC at first MQTT Connection (Security), and set pseudocode operation for publish Data and Get Firmware (To Connect with also implemented procedure)

+++ platformio/stima_v4/libraries/RmapCommon/typedef.h
@@ -40,7 +40,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
 #define TYPE_LENGTH (5)
 
-typedef int32_t rmapdata_t;
+typedef uint16_t rmapdata_t;
 
 #define RMAPDATA_MAX            (std::numeric_limits<rmapdata_t>::max())
 #define RMAPDATA_MIN            (std::numeric_limits<rmapdata_t>::min())
